Исправляет структуру проекта и корректирует компоненты
