+++ my-app/src/components/Chat/Chat.css
@@ -1,0 +1,39 @@
+.chat__wrapper {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  flex-shrink: 0;
+  width: 100%;
+  max-width: 1200px;
+  margin: -165px auto 42px;
+  border-radius: 4px;
+  background-color: #ffffff;
+  box-shadow: 0px 1px 4px 0px rgba(0, 0, 0, 0.3);
+}
+
+@media screen and (max-height: 900px) {
+  .chat__wrapper {
+    margin-bottom: 0;
+  }
+}
+
+/* Let's get this party started */
+::-webkit-scrollbar {
+    width: 6px;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+    /*-webkit-box-shadow: inset 0 0 2px rgba(0,0,0,0.15);*/
+    -webkit-border-radius: 0px;
+    border-radius: 0px;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+    -webkit-border-radius: 2px;
+    border-radius: 2px;
+    background: rgba(0, 0, 0, 0.05);
+    -webkit-box-shadow: inset 0 0 2px rgba(0,0,0,0.13);
+}
+
