+++ my-app/src/components/ContentChat/ContentChat.css
@@ -1,0 +1,26 @@
+.chat__messages {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex-grow: 1;
+  flex-shrink: 1;
+  flex-basis: 100%;
+  max-height: 600px;
+  min-height: 280px;
+  margin: 0;
+  padding: 28px 42px 28px 48px;
+  overflow-y: auto;
+}
+
+@media (min-width: 320px) and (max-width: 600px) {
+  .chat__messages {
+    padding-left: 0;
+    padding-right: 0;
+  }
+}
+
+@media screen and (max-height: 900px) {
+  .chat__messages {
+    height: 419px;
+  }
+}
