+++ my-app/src/components/HeaderChat/HeaderChat.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 
+import './HeaderChat.css';
+
 
 class HeaderChat extends Component {
   render() {
