+++ my-app/src/components/SendMessageForm/SendMessageForm.css
@@ -1,0 +1,90 @@
+.chat__nav {
+  flex-grow: 0;
+  flex-shrink: 0;
+  margin-top: auto;
+  padding: 32px 48px;
+  background-color: #f1f1f1;
+  border-bottom-left-radius: 3px;
+  border-bottom-right-radius: 3px;
+}
+
+.chat-form {
+  display: flex;
+  width: 100%;
+}
+
+.chat-form__field {
+  flex-grow: 1;
+  flex-shrink: 1;
+  min-height: 74px;
+  margin-right: 32px;
+  padding: 16px 24px;
+  font: 300 16px/1.25 "Roboto", "Arial", sans-serif;
+  color: #434343;
+  border-radius: 2px;
+  border: 1px solid #e5e5e5;
+  outline: none;
+  transition: border-color 0.15s ease-in-out;
+}
+
+.chat-form__field::placeholder {
+  color: rgba(0, 0, 0, 0.3);
+}
+
+.chat-form__field:hover {
+  border-color: #dcdbdb;
+}
+
+.chat-form__field:focus {
+  border-color: #ffb4ce;
+}
+
+.chat-form__btn {
+  align-self: center;
+  display: inline-block;
+  box-sizing: border-box;
+  width: 150px;
+  height: 40px;
+  padding: 0 30px;
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 40px;
+  vertical-align: middle;
+  text-align: center;
+  text-decoration: none;
+  word-wrap: break-word;
+  color: #ffffff;
+  background-color: #ff4081;
+  border-radius: 2px;
+  border: 0;
+  box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.24);
+  user-select: none;
+  outline: none;
+  cursor: pointer;
+  transition: box-shadow 0.15s ease-in-out;
+}
+
+.chat-form__btn:hover,
+.chat-form__btn:active {
+  box-shadow: 0 3px 5px 0 rgba(0, 0, 0, 0.16);
+}
+
+@media (min-width: 320px) and (max-width: 600px) {
+  .chat__nav {
+    padding-left: 24px;
+    padding-right: 24px;
+  }
+
+  .chat-form {
+    flex-direction: column;
+  }
+
+  .chat-form__field {
+    margin-right: 0;
+    margin-bottom: 24px;
+  }
+
+  .chat-form__btn {
+    margin-right: auto;
+  }
+}
