+++ my-app/src/components/SendMessageForm/SendMessageForm.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 
+import './SendMessageForm.css';
+
 
 class SendMessageForm extends Component {
   render() {
